myapp/requirements.txt
crcmod==1.7
cryptography==41.0.3
cssselect==1.2.0
dataclasses-json==0.5.14
distro==1.8.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.3
firebase-admin==6.2.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-kms==2.19.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
google-search-results==2.4.2
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.5
langchain-core==0.1.2
langsmith==0.0.72
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
numexpr==2.8.6
numpy==1.25.2
openai==1.6.0
packaging==23.2
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
rsa==4.9
sgmllib3k==1.0.0
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
tenacity==8.2.3
tinysegmenter==0.3
tldextract==3.5.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
